# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flame Sales & Expense Tracker API",
    "version": "1.0.0",
    "description": "API for managing organizations, projects, expenses, sales, and related resources in the Flame Sales & Expense Tracker."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://flame-expense-tracker.bagumajonah3.workers.dev",
      "description": "Production server (Cloudflare Workers)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Flame API key in the format flame_ak_<prefix>_<secret>. Include it as an X-Api-Key header when calling these endpoints."
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/reports/pnl-by-project": {
      "get": {
        "operationId": "getPnlByProject",
        "tags": [
          "Reports"
        ],
        "summary": "Get profit and loss by project",
        "description": "Returns total revenue, total expenses, and net profit per project for the current organization.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cycleId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "P&L data by project."
          },
          "401": {
            "description": "API key required."
          }
        }
      }
    },
...
  }
}